+++ Sources/Switchboard/SwitchboardSupportedProvider.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+
+public struct SupportedProvider: Sendable, Hashable {
+    public let id: String
+
+    public init(_ id: String) {
+        self.id = id
+    }
+}